requirements.txt
python-telegram-bot[job-queue]==20.8
sqlalchemy==2.0.25
spotipy==2.23.0
requests==2.31.0
python-dotenv==1.0.0
yt-dlp==2024.3.10
apscheduler==3.10.4
pytz==2024.1
aiohttp==3.9.3
psycopg2-binary==2.9.9